ed25519_compact

Struct SecretKey

pub struct SecretKey(/* private fields */);

A secret key.

Implementations

impl SecretKey

pub const BYTES: usize

Number of bytes in a secret key.

pub fn new(sk: [u8; 64]) -> Self

Creates a secret key from raw bytes.

pub fn from_slice(sk: &[u8]) -> Result<Self, Error>

Creates a secret key from a slice.

pub fn public_key(&self) -> PublicKey

Returns the public counterpart of a secret key.

pub fn seed(&self) -> Seed

Returns the seed of a secret key.

pub fn validate_public_key(&self, pk: &PublicKey) -> Result<(), Error>

Returns Ok(()) if the given public key is the public counterpart of this secret key. Returns Err(Error::InvalidPublicKey) otherwise. The public key is recomputed (not just copied) from the secret key, so this will detect corruption of the secret key.

impl SecretKey

pub fn sign_incremental(&self, noise: Noise) -> SigningState

Sign a multi-part message (streaming API). It is critical for noise to never repeat.

pub fn sign(&self, message: impl AsRef<[u8]>, noise: Option<Noise>) -> Signature

Computes a signature for the message message using the secret key. The noise parameter is optional, but recommended in order to mitigate fault attacks.

impl SecretKey

pub fn from_der(der: &[u8]) -> Result<Self, Error>

Import a secret key from an OpenSSL-compatible DER file.

pub fn from_pem(pem: &str) -> Result<Self, Error>

Import a secret key from an OpenSSL-compatible PEM file.

pub fn to_der(&self) -> Vec<u8>

Export a secret key as an OpenSSL-compatible DER file.

pub fn to_pem(&self) -> String

Export a secret key as an OpenSSL-compatible PEM file.